change sales rep level

+++ database/migrations/2020_11_25_113900_add_sales_rep_column_sales_rep_fee_to_customers_table.php
@@ -14,7 +14,7 @@ class AddSalesRepColumnSalesRepFeeToCustomersTable extends Migration
     public function up()
     {
         Schema::table('customers', function (Blueprint $table) {
-            $table->integer('sales_rep_fee')->nullable();
+            $table->decimal('sales_rep_fee')->nullable();
             $table->unsignedBigInteger('sales_rep_id')->nullable();
             $table->foreign('sales_rep_id')
                     ->references('id')
